backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
fpdf2
python-dotenv
pandas
numpy
joblib
scikit-learn
qrcode[pil]
requests
Pillow
httpx
